backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.11.1
pydantic>=2.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
transformers>=4.30.2
torch>=2.0.1
pandas>=2.0.2
numpy>=1.24.3
scikit-learn>=1.2.2
spacy>=3.5.3
kafka-python>=2.0.2
pytest>=7.3.1
httpx>=0.24.1
pillow>=9.5.0
python-dotenv>=1.0.0
requests>=2.31.0
tweepy>=4.14.0
instaloader>=4.9.5
praw>=7.7.0